% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEstFormat.R
\name{addEstFormat}
\alias{addEstFormat}
\title{Add formatted columns to a paremeter table}
\usage{
addEstFormat(pars, source.ci = "cov", rse.cov)
}
\arguments{
\item{pars}{Parameter estimates (a data.frame, as generated by
createParameterTable).}

\item{source.ci}{Source of confidence intervals. "cov" for "Covariance step", "boot" for bootstrap, and "" for nothing.}

\item{rse.cov}{Include relative standard error estimate based on covariate step? Default is TRUE if source=="cov" and FALSE if not.}
}
\description{
An internal function to add character label columns depending on parameter
type and transformation.
}
