% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameterTable.R
\name{createParameterTable}
\alias{createParameterTable}
\title{Create parameter table data.frame using `NMreadExt()` and
`NMreadParsText()`}
\usage{
createParameterTable(file.lst, args.ParsText, dt.labels = NULL, by.labels)
}
\arguments{
\item{file.lst}{Control stream. If possible, it is recommended to
use output control stream. You can also use input control
stream.}

\item{args.ParsText}{List of arguments to be passes to
`NMreadParsText()`.}

\item{dt.labels}{Optional table of labels to overwrite results
from `NMreadParsText()`. Useful to correcissues on a model you
don´t want to rerun.}

\item{by.labels}{If `dt.labels` provided, names of columns to
merge by.}
}
\description{
Create parameter table data.frame using `NMreadExt()` and
`NMreadParsText()`
}
\seealso{
formatParTable
}
