% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModels.R
\name{findModels}
\alias{findModels}
\title{find model paths winin a directory}
\usage{
findModels(
  dir,
  pattern = ".*.lst$",
  methods = c("main", "bbr"),
  recursive = FALSE,
  ext.mod.bbr = ".mod"
)
}
\arguments{
\item{methods}{Prioritized vector of model types to search
for. "main" means psn and nmsim because they copy results to
main dir.}
}
\description{
Finds models created by NMsim, BBR, and PSN.
}
\details{
"main": any lst in main dir "bbr": any lst generated this
    way: look for x.mod files, look for x/ and for x/x.lst. The
    x.lst is the result. But a x.mod is used to compare to results
    from "main". When x.lst (derived from x.mod) matches a lst
    file from main, the order of methods is used to prioritize.
}
