% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatParameterTable.R
\name{formatParameterTable}
\alias{formatParameterTable}
\title{Deprecated: Quickly format a parameter table into flextable objects}
\usage{
formatParameterTable(
  pars,
  include.fix = "ifNotZero",
  include,
  drop,
  drop.symbol = "not used"
)
}
\arguments{
\item{pars}{A parameter table as generated by `NMwork::createParameterTable()`.}

\item{include.fix}{Either a logical or "ifNotZero".}

\item{include}{See code}

\item{drop}{See code}

\item{drop.symbol}{See code}
}
\description{
Use createParameterTable() and printParameterTable() instead.
}
\details{
A simple printing function for createParameterTable(). It creates three objects.
\itemize{
\item partab_full: a df with all columns for selected parameters.
\item partab_detail: a flextable with selected columns for selected parameters.
\item partab_present: a flextable with fewer selected columns for selected parameters.
}

Notes: I suggest dropping the asterisk on parameter.ltx and
instead include a nicely formated transformation column in
createParameterTable(). It could be included in a parameter table
only if non-standard transformations are used.

formatParameterTable is a bad name. This function selects rows and
columns in three different levels of detail and returns those
three objects. Formatting should happen in
`createParameterTable()`.
}
