% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSource.R
\name{getSource}
\alias{getSource}
\title{Get and source an R script.}
\usage{
getSource(
  file,
  dir.central = NULL,
  dir.local,
  overwrite = FALSE,
  source.directly = FALSE,
  quiet = T
)
}
\arguments{
\item{file}{File name of wanted file}

\item{dir.central}{Folder path of wanted file}

\item{dir.local}{Where to put the file if imported, and where to look for
already imported files. Default is getwd().}

\item{overwrite}{Owerwrite previously imported file?}

\item{source.directly}{Enables direct sourcing of the central file copy. This
bypasses the whole concept of the function but it is useful when
developing while using a function. Especially if your debugger in the
editor is linking to a file that you will edit while debugging. It gives
a warning because it is not recommended in final code.}

\item{quiet}{Disables printning. Mainly used in testing.}
}
\value{
None. Sources the specified file into the global environment.
}
\description{
Get file from remote library and inport it into project if it does
not exist. Then source it.
}
\concept{FileSystem}
