% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLabel.R
\name{mergeLabel}
\alias{mergeLabel}
\title{Add parameter labels from a parameter table}
\usage{
mergeLabel(
  pars,
  tab.label = NULL,
  by = "parameter",
  col.label = "parameter",
  suffix = NULL
)
}
\arguments{
\item{pars}{parameters table}

\item{tab.label}{table with a column for labels}

\item{by}{column to merge pars and tab.label}

\item{col.label}{column in `tab.label` to use for labelling}

\item{suffix}{not sure}
}
\description{
Add parameter labels from a parameter table
}
\details{
This is internally used in plotEstCor and plotTrace. Not
    sure it should its worth exporting.
}
