% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newModel.R
\name{newModel}
\alias{newModel}
\title{Create new control stream based on an existing model}
\usage{
newModel(
  file.mod,
  newfile,
  update = TRUE,
  values,
  description = NULL,
  based.on,
  author = NULL,
  write.file = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{If newfile exists, overwrite it?}
}
\description{
Create new control stream based on an existing model
}
