% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMscript.R
\name{NMscript}
\alias{NMscript}
\title{Run script on models using Rscript, optionally using sge}
\usage{
NMscript(
  script,
  file.mod,
  ...,
  sge = FALSE,
  wait = FALSE,
  nc = 4,
  stdout_file,
  stderr_file
)
}
\arguments{
\item{script}{The actual command-line command you want to run on
the grid (including all arguments)}

\item{file.mod}{First argument sent to script. If length>1, the
script is run on all the elements.}

\item{...}{Additional arguments to script.}

\item{sge}{Send to cluster using `qsub`?}

\item{wait}{Only used if sge is `FALSE`.}

\item{nc}{Only used if sge is `TRUE`. The number of cores you want
the script to run on the grid with.}

\item{stdout_file}{Only used if sge is `TRUE`. The file where the
job output will be printed.}

\item{stderr_file}{Only used if sge is `TRUE`. The file where the
job errror output will be printed.}
}
\description{
Run script on models using Rscript, optionally using sge
}
