% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc.R
\name{plot_vpc}
\alias{plot_vpc}
\title{Plot VPC data generated by tidyvpc using custom function}
\usage{
plot_vpc(
  .vpcdata = vpc.predcorr.all,
  .predcorr = FALSE,
  .pointsize = 0.6,
  .subset = list(Study = "XXX-XXX-XXX"),
  .x_axis_lab = "Time after first dose (h)",
  .y_axis_lab =
    "Percentiles and associated 95\% CI\\nprediction-corrected concentration (ng/mL)"
)
}
\arguments{
\item{.vpcdata}{dataset generated by tidyvpc}

\item{.predcorr}{TRUE/FALSE - whether to use prediction-corrected observations or not}

\item{.pointsize}{size of points for observations in plot}

\item{.subset}{subset of data?}

\item{.x_axis_lab}{x-axis label}

\item{.y_axis_lab}{y-axis label}
}
\description{
Plot VPC data generated by tidyvpc using custom function
}
