% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstCor.R
\name{plotEstCor}
\alias{plotEstCor}
\title{Plot correlation matrix heatmap from completed NONMEM run}
\usage{
plotEstCor(file.lst, pars = NULL, label.by = "parameter", col.label = NULL)
}
\arguments{
\item{file.lst}{`.lst` file from completed NONMEM run}

\item{pars}{a table of parameter names and labels to use for renaming parameters to descriptive labels}
}
\description{
Plot correlation matrix heatmap from completed NONMEM run
}
