% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndivProfilesDvPredIpred.R
\name{plotIndivProfilesDvPredIpred}
\alias{plotIndivProfilesDvPredIpred}
\title{Plot Inidividual profiles observed (DV) and predicted (PRED and IPRED)}
\usage{
plotIndivProfilesDvPredIpred(
  .data = NULL,
  .timecol = "TIME",
  .dvcol = "DV",
  .predcol = "PRED",
  .ipredcol = "IPRED",
  .log = FALSE
)
}
\arguments{
\item{.data}{dataset to use for plotting}

\item{.timecol}{column to use for x-axis (TIME)}

\item{.dvcol}{column to use for y-axis for observations (DV)}

\item{.predcol}{column to use for y-axis for population predictions (PRED)}

\item{.ipredcol}{column to use for y-axis for individual predictions (PRED)}

\item{.log}{should we use a logged y-axis?}
}
\description{
Plot Inidividual profiles observed (DV) and predicted (PRED and IPRED)
}
