% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResPred.R
\name{plotResPred}
\alias{plotResPred}
\title{Plot CWRES/IWRES vs population predictions (PRED)}
\usage{
plotResPred(
  .data = NULL,
  .file.mod = NULL,
  .ResCol = "CWRES",
  .PredCol = "PRED",
  .AddSmooth = FALSE
)
}
\arguments{
\item{.data}{dataset object to use for plotting. Must have all columns
required for plotting}

\item{.file.mod}{the path to a completed NONMEM model which will be passed
to `NMdata::NMscanData()` to read in the dataset}

\item{.ResCol}{the residuals column name (i.e. "RES", "CWRES", "NPDE",
etc.). Will be plotted as the y-axis.}

\item{.PredCol}{the population prediction column name (i.e. "PRED"). Will be plotted as the x-axis.}

\item{.AddSmooth}{TRUE/FALSE. If TRUE, will add a smooth line using method="loess"}
}
\description{
Plot CWRES/IWRES vs population predictions (PRED)
}
