% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{plotTrace}
\alias{plotTrace}
\title{Plot parameter estimation iterations}
\usage{
plotTrace(file.mod, pars = NULL, label.by = "parameter", col.label = NULL, ...)
}
\arguments{
\item{...}{Passed to `NMdata::NMreadExt()`.}
}
\description{
Plot parameter estimation iterations
}
