% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printParameterTable.R
\name{printParameterTable}
\alias{printParameterTable}
\title{Print parameter table to console, flextable, latex, or (latex) pdf file}
\usage{
printParameterTable(
  pars,
  engine = "kable",
  format,
  footnotes = NULL,
  script = NULL,
  file.mod,
  include,
  include.pattern,
  drop,
  drop.pattern,
  include.fix,
  caption,
  ci = "cov",
  rse.cov = TRUE,
  label.pmtables
)
}
\arguments{
\item{pars}{A parameter table as generated by
`NMwork::createParameterTable()`.}

\item{engine}{"kable", "pmtables" or "flextable".}

\item{format}{A format like "tex" or "pdf" for a preview or a file name for
saving a file. What formats are available depends on `engine`.}

\item{footnotes}{Additional footnotes to include. Footnotes about theta
transformations, abbreviations, and source paths will automatically be
included.}

\item{script}{Path to source code to include in source code.}

\item{file.mod}{Path to Nonmem control stream. Normally when `pars` was
created with `createParameterTable()`, this is not needed because `pars`
will contain this information already. If so, file.mod should only be
used in cases where modifications to the original parameter table are
being performed to generate new ones.}

\item{include}{A character vector of symbol values to include. All else will
be dropped.}

\item{include.pattern}{A regular expression run on the `symbol` column. What
matches will be kept.}

\item{drop}{A character vector of symbol values to drop. All else will be
kept.}

\item{drop.pattern}{A regular expression run on the `symbol` column. What
matches will be dropped.}

\item{include.fix}{Either a logical or "NotZero". `TRUE` means fixed
parameters are not filtered out, `FALSE` means they are filtered out,
"NotZero" means they are only kept if not equal to zero.}

\item{caption}{If supplied, this text will be added to a simple model naming
string, based on the `file.mod` file name.}

\item{ci}{The default ("cov") is to attempt to extract parameter estimation
uncertainty from a succesful covariate step. To use a bootstrap estimate
instead,}

\item{rse.cov}{Include relative standard error estimate based on covariate
step? Default is TRUE if source=="cov" and FALSE if not.}

\item{label.pmtables}{If `engine=="pmtables"`, a latex label is needed. Like
label.pmtables="tab:myfirsttable".}
}
\description{
Print parameter table to console, flextable, latex, or (latex) pdf file
}
\details{
kable: for rmd and rmd-latex: can't get longtable to work with
    footnotes. Currently not using footnotes.

pmtables is only available if the package called pmtables is
installed. pmtables is not available on CRAN, at the time of writing
this. It can be installed from MPN.
}
\examples{
file.mod <- system.file("nonmem/xgxr134.mod",package="NMwork")
pars <- createParameterTable(file.lst=file.mod,
                  args.ParsText=list(format="\%idx; \%symbol ; \%label [\%unit] ; \%trans",
                                     format.sigma="\%symbol - \%label "))
printParameterTable(pars=pars,format="R")
printParameterTable(pars=pars,format="rmd")
}
