% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cor_mat.R
\name{read_cor_mat}
\alias{read_cor_mat}
\title{read list of correlation matrices from .cor file. takes the path to the .cor
file as an argument. output is a list with length equal to the number of
tables output in the .cor file. Use: when more than one table is listed in
.cor file and you want all of them}
\usage{
read_cor_mat(file.lst, tableno = "max")
}
\arguments{
\item{file.lst}{NONMEM model output file, must be completed run}

\item{tableno}{which correlation matrix to output in case there are
multiple. tableno must be either one of the character strings min, max,
all or an integer greater than zero}
}
\description{
read list of correlation matrices from .cor file. takes the path to the .cor
file as an argument. output is a list with length equal to the number of
tables output in the .cor file. Use: when more than one table is listed in
.cor file and you want all of them
}
