% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptPaths.R
\name{scriptPaths}
\alias{scriptPaths}
\title{Create list of script paths and other info related to a script}
\usage{
scriptPaths(path)
}
\arguments{
\item{path}{Absolute path to script}
}
\value{
A list of
\itemize{
\item path Absolute path "/data/dir/script.R"
\item name "script"
\item label "dir/script.R"
}
}
\description{
Create list of script paths and other info related to a script
}
