% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapTextToLength.R
\name{wrapTextToLength}
\alias{wrapTextToLength}
\title{Does a script fit on a single page?}
\usage{
wrapTextToLength(
  text =
    "Source code: path/to/some/file/thatIsVery/Very/Veryvery/looooooooonnnnng/scripts/testing.R",
  page_width = 6.26,
  font_size = 7,
  font_family = "sans"
)
}
\arguments{
\item{text}{string to be wrapped}

\item{page_width}{}

\item{font_size}{}

\item{font_family}{}
}
\description{
Take script path and test if it is too long to fit on a typical
page. If it is too long, wrap it onto multiple lines. This is more
involved than one may think as shown here.
}
