% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCoal.R
\name{mergeCoal}
\alias{mergeCoal}
\title{Overwrite values in one data.frame if they are available in another.}
\usage{
mergeCoal(x, y, by, cols.coal, as.fun)
}
\arguments{
\item{x}{The initial data.frame}

\item{y}{A data.frame to prioritize overc `x`.}

\item{by}{Columns to merge by}

\item{cols.coal}{Columns to overwrite values from `y` if available.}
}
\description{
Repair x with y
}
\details{
Non-na values in y will be used o overwrite columns in x
    at the rows matched using `by` columns.

Merges must be done using the same "by" columns for all rows. If
rows needs to be merged using varying by columns, the merges must
be done sequentially.
}
\keyword{internal}
