% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPaths.R
\name{modelPaths}
\alias{modelPaths}
\title{Create a convenient set of file paths and other info related to a model.}
\usage{
modelPaths(file, as.dt = FALSE, simplify = TRUE)
}
\arguments{
\item{file}{path to control stream, or almost any other model file}

\item{as.dt}{Return model information in a data.table?}

\item{simplify}{If only one model supplied, and `as.dt=FALSE`,
should the format be simplified to a list of model elements,
rather than a list of (model) lists of elements?}
}
\description{
Create a convenient set of file paths and other info related to a model.
}
