% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDVpredIpred.R
\name{plotDVpredIpred}
\alias{plotDVpredIpred}
\title{Plot DV vs PRED or IPRED}
\usage{
plotDVpredIpred(
  .data = NULL,
  .file.mod = NULL,
  .dvcol = "DV",
  .predcol = "PRED",
  .log = FALSE
)
}
\description{
Plot DV vs PRED or IPRED
}
