% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{plotTrace}
\alias{plotTrace}
\title{Plot parameter estimation iterations}
\usage{
plotTrace(file.mod, pars = NULL, label.by = "parameter", col.label = NULL, ...)
}
\arguments{
\item{file.mod}{NONMEM model file, must be completed run}

\item{pars}{a parameter table to map NONMEM parameters to their labels}

\item{label.by}{a column in the parameter table which will be used for merging the parameter table and the parameter labels in the .ext file, as read by `NMdata::NMreadExt()`. usually this column is 'parameter' and has values such as 'THETA1', 'THETA2', .... , 'OMEGA(1,1)', 'OMEGA(2,2)', ... 'SIGMA(1,1)',  etc.}

\item{col.label}{the column in the parameter table which we will use for labeling the plot panels.}

\item{...}{Passed to `NMdata::NMreadExt()`.}
}
\description{
Plot parameter estimation iterations
}
