% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printParameterTable.R
\name{printParameterTable}
\alias{printParameterTable}
\title{Print parameter table to console, flextable, latex, or (latex) pdf file}
\usage{
printParameterTable(
  pars,
  format,
  footnotes = NULL,
  script = NULL,
  file.mod,
  include,
  include.pattern,
  drop,
  drop.pattern,
  include.fix
)
}
\arguments{
\item{pars}{A parameter table as generated by `NMwork::createParameterTable()`.}

\item{include}{A character vector of symbol values to include. All else will be dropped.}

\item{include.pattern}{A regular expression run on the `symbol` column. What matches will be kept.}

\item{drop}{A character vector of symbol values to drop. All else will be kept.}

\item{drop.pattern}{A regular expression run on the `symbol` column. What matches will be dropped.}

\item{include.fix}{Either a logical or "ifNotZero". `TRUE` means fixed parameters are not filtered out, `FALSE` means they are filtered out, "ifNotZero" means they are only kept if not equal to zero.}
}
\description{
Print parameter table to console, flextable, latex, or (latex) pdf file
}
\examples{
file.mod <- system.file("nonmem/xgxr134.mod",package="NMwork")
pars <- createParameterTable(file.lst=file.mod,
                  args.ParsText=list(format="\%idx; \%symbol ; \%label [\%unit] ; \%trans",
                                     format.sigma="\%symbol - \%label "))
printParameterTable(pars=pars,format="R")
printParameterTable(pars=pars,format="rmd")
}
